# lib/api/generated/model/error.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * Fitness App API
 * API completa para o sistema de fitness com autenticação, gerenciamento de usuários e muito mais
 * OpenAPI spec version: 1.0.0
 */

export interface Error {
  /** Mensagem de erro */
  error?: string;
  /** Detalhes do erro (apenas em desenvolvimento) */
  details?: string;
}