# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fitness App API",
    "version": "1.0.0",
    "description": "API completa para o sistema de fitness com autenticação, gerenciamento de usuários e muito mais",
    "contact": {
      "name": "Suporte API",
      "email": "suporte@fitnessapp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro"
          },
          "details": {
            "type": "string",
            "description": "Detalhes do erro (apenas em desenvolvimento)"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "name": {
            "type": "string",
            "description": "Nome completo do usuário"
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "GYM",
              "ADMIN"
            ],
            "description": "Role do usuário (fonte da verdade)"
          },
          "hasGym": {
            "type": "boolean",
            "description": "Se o usuário tem perfil de academia"
          },
          "hasStudent": {
            "type": "boolean",
            "description": "Se o usuário tem perfil de aluno"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do usuário (ISO 8601)",
            "example": "2025-12-27T23:56:23.743Z"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID da sessão"
          },
          "token": {
            "type": "string",
            "description": "Token de sessão"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Não autenticado ou sessão inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Não autenticado"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/exercises/search": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Buscar exercícios",
        "description": "Busca exercícios do banco educacional com filtros opcionais",
        "operationId": "searchExercises",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Termo de busca (nome do exercício)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "muscle",
            "in": "query",
            "description": "Filtrar por grupo muscular",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de exercícios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exercises": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "primaryMuscles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "group": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/api/auth/session": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Verificar sessão",
        "description": "Retorna informações sobre a sessão atual do usuário autenticado.",
        "operationId": "getSession",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sessão válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "user": {
                    "id": "clx1234567890",
                    "email": "joao@email.com",
                    "name": "João Silva",
                    "role": "STUDENT",
                    "hasGym": false,
                    "hasStudent": true,
                    "createdAt": "2025-12-27T23:56:23.743Z"
                  },
                  "session": {
                    "id": "sess_123",
                    "token": "session-1234567890-abc123"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou sessão inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "noToken": {
                    "value": {
                      "error": "Token não fornecido"
                    }
                  },
                  "invalidSession": {
                    "value": {
                      "error": "Sessão inválida ou expirada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}